secure-data-sharing-main/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
passlib[bcrypt]
python-jose[cryptography]
cryptography
pandas
numpy
scikit-learn
email-validator
